+++ mongodb-sync-v4-driver/src/main/java/com/github/cloudyrock/mongock/driver/mongodb/sync/v4/driver/MongoSync4Driver.java
@@ -1,21 +1,55 @@
 package com.github.cloudyrock.mongock.driver.mongodb.sync.v4.driver;
 
 import com.github.cloudyrock.mongock.driver.api.entry.ChangeEntry;
-import com.github.cloudyrock.mongock.driver.api.entry.ChangeEntryService;
-import com.github.cloudyrock.mongock.driver.mongodb.sync.v4.changelogs.runalways.MongockSync4LegacyMigrationChangeRunAlwaysLog;
-import com.github.cloudyrock.mongock.driver.mongodb.sync.v4.changelogs.runonce.MongockSync4LegacyMigrationChangeLog;
-import com.github.cloudyrock.mongock.driver.mongodb.sync.v4.repository.MongoSync4ChangeEntryRepository;
+import com.github.cloudyrock.mongock.exception.MongockException;
 import com.github.cloudyrock.mongock.utils.TimeService;
 import com.github.cloudyrock.mongock.utils.annotation.NotThreadSafe;
+import com.mongodb.MongoClientException;
+import com.mongodb.client.ClientSession;
 import com.mongodb.client.MongoClient;
-import com.mongodb.client.MongoDatabase;
+import com.mongodb.client.TransactionBody;
 
 @NotThreadSafe
 public class MongoSync4Driver extends MongoSync4DriverBase<ChangeEntry> {
 
-  protected MongoSync4ChangeEntryRepository<ChangeEntry> changeEntryRepository;
+  private MongoClient mongoClient;
 
-  private static final TimeService TIME_SERVICE = new TimeService();
+  protected MongoSync4Driver(MongoClient mongoClient,
+                             String databaseName,
+                             long lockAcquiredForMillis,
+                             long lockQuitTryingAfterMillis,
+                             long lockTryFrequencyMillis) {
+    super(mongoClient.getDatabase(databaseName), lockAcquiredForMillis, lockQuitTryingAfterMillis, lockTryFrequencyMillis);
+    this.mongoClient = mongoClient;
+  }
+
+  @Override
+  public void executeInTransaction(Runnable operation) {
+    ClientSession clientSession;
+    try {
+      clientSession = mongoClient.startSession();
+    } catch (MongoClientException ex) {
+      throw new MongockException("ERROR starting session. If Mongock is connected to a MongoDB cluster which doesn't support transactions, you must to disable transactions", ex);
+    }
+    try {
+      clientSession.withTransaction(getTransactionBody(operation), txOptions);
+    } catch (Exception ex) {
+      throw new MongockException(ex);
+    } finally {
+      clientSession.close();
+    }
+  }
+
+  private TransactionBody getTransactionBody(Runnable operation) {
+    return (TransactionBody<String>) () -> {
+      operation.run();
+      return "Mongock transaction operation";
+    };
+  }
+
+  ////////////////////////////////////////////////////////////
+  //BUILDER METHODS
+  ////////////////////////////////////////////////////////////
 
 
   //TODO CENTRALIZE DEFAULT PROPERTIES
@@ -24,17 +58,16 @@ public class MongoSync4Driver extends MongoSync4DriverBase<ChangeEntry> {
   }
 
   public static MongoSync4Driver withLockStrategy(MongoClient mongoClient,
-                                                 String databaseName,
-                                                 long lockAcquiredForMillis,
-                                                 long lockQuitTryingAfterMillis,
-                                                 long lockTryFrequencyMillis) {
+                                                  String databaseName,
+                                                  long lockAcquiredForMillis,
+                                                  long lockQuitTryingAfterMillis,
+                                                  long lockTryFrequencyMillis) {
     return new MongoSync4Driver(mongoClient, databaseName, lockAcquiredForMillis, lockQuitTryingAfterMillis, lockTryFrequencyMillis);
   }
 
 
   /**
-   * @Deprecated
-   * Use withLockStrategy instead
+   * @Deprecated Use withLockStrategy instead
    */
   @Deprecated
   public static MongoSync4Driver withLockSetting(MongoClient mongoClient,
@@ -42,38 +75,10 @@ public class MongoSync4Driver extends MongoSync4DriverBase<ChangeEntry> {
                                                  long lockAcquiredForMinutes,
                                                  long maxWaitingForLockMinutes,
                                                  int maxTries) {
-    long lockAcquiredForMillis = TIME_SERVICE.minutesToMillis(lockAcquiredForMinutes);
-    long lockQuitTryingAfterMillis = TIME_SERVICE.minutesToMillis(maxWaitingForLockMinutes * maxTries);
+    TimeService timeService = new TimeService();
+    long lockAcquiredForMillis = timeService.minutesToMillis(lockAcquiredForMinutes);
+    long lockQuitTryingAfterMillis = timeService.minutesToMillis(maxWaitingForLockMinutes * maxTries);
     long tryFrequency = 1000L;// 1 second
     return MongoSync4Driver.withLockStrategy(mongoClient, databaseName, lockAcquiredForMillis, lockQuitTryingAfterMillis, tryFrequency);
-  }
-
-  // For children classes like SpringData drivers
-  protected MongoSync4Driver(MongoDatabase mongoDatabase,
-                             long lockAcquiredForMillis,
-                             long lockQuitTryingAfterMillis,
-                             long lockTryFrequencyMillis) {
-    super(mongoDatabase, lockAcquiredForMillis, lockQuitTryingAfterMillis, lockTryFrequencyMillis);
-  }
-
-  protected MongoSync4Driver(MongoClient mongoClient,
-                             String databaseName,
-                             long lockAcquiredForMillis,
-                             long lockQuitTryingAfterMillis,
-                             long lockTryFrequencyMillis) {
-    super(mongoClient, databaseName, lockAcquiredForMillis, lockQuitTryingAfterMillis, lockTryFrequencyMillis);
-  }
-
-  @Override
-  public ChangeEntryService<ChangeEntry> getChangeEntryService() {
-    if (changeEntryRepository == null) {
-      this.changeEntryRepository = new MongoSync4ChangeEntryRepository<>(mongoDatabase.getCollection(changeLogCollectionName), indexCreation, getReadWriteConfiguration());
-    }
-    return changeEntryRepository;
-  }
-
-  @Override
-  public Class getLegacyMigrationChangeLogClass(boolean runAlways) {
-    return runAlways ? MongockSync4LegacyMigrationChangeRunAlwaysLog.class : MongockSync4LegacyMigrationChangeLog.class;
   }
 }
